Fix compatibility for EL5 and EL6 systems for AuthorizedKeysCommandRunAs vs. -AuthorizedKeysCommandUser
It seems the AuthorizedKeysCommand was added to OpenSSH in version 6.2. It also seems RedHat implemented the same thing in the version of SSH they shipped with versions 5 and 6. However, when the feature was added to OpenSSH the configuration key to specify the user to run the command as was changed from what RedHat had implemented.